Modify Translator to return BaseObject by default, not None.
Fixes #31.
