--- test/unit/object/test_translator.py
@@ -1,3 +1,0 @@
-# coding: utf-8
-
-from __future__ import unicode_literals
